extract plots to their own classes
